type: feat
A new model was added. This model have a batch norm layer on the 4th conv layer, weight initialisation as previously and 600 epochs to train. The loss is lower than the previous model.
However, when visualizing keypoints during inference, the model 4 seems to be better than this model.
Next improvement :
	- Add Batch norm after each conv & Linear model
	- Run for 800-900 epochs